Update ins_forward_simulation.R
Took out a line, according to Zita, so that the simulations don't given an Option 3 (i.e., we will only have no choice, Option 1, Option 2, which is consistent to what's set up in VBA). How did you want to update the upper/lower bound of nu, after the sign change; should we make it entirely unbounded?